fix: Route checker
